network: move context declaration to private scope
