+++ include/fluent-bit/flb_network.h
@@ -26,7 +26,6 @@
 #include <fluent-bit/flb_uri.h>
 #include <fluent-bit/flb_upstream_conn.h>
 #include <fluent-bit/flb_net_dns.h>
-#include <ares.h>
 
 #define FLB_NETWORK_DEFAULT_BACKLOG_SIZE              128
 #define FLB_NETWORK_UNIX_SOCKET_PEER_ADDRESS_TEMPLATE "pid_%s"
@@ -108,31 +107,6 @@ struct flb_net_host {
     flb_sds_t listen;      /* Listen interface       */
     struct flb_uri *uri;   /* Extra URI parameters   */
 };
-
-/* Defines an async DNS lookup context */
-struct flb_dns_lookup_context {
-    struct mk_event              response_event;                  /* c-ares socket event */
-    int                          ares_socket_registered;
-    struct ares_socket_functions ares_socket_functions;
-    int                         *udp_timeout_detected;
-    int                          ares_socket_created;
-    int                          ares_socket_type;
-    void                        *ares_channel;
-    int                         *result_code;
-    struct mk_event_loop        *event_loop;
-    struct flb_coro             *coroutine;
-    struct flb_sched_timer      *udp_timer;
-    int                          finished;
-    int                          dropped;
-    struct flb_net_dns          *dns_ctx;
-    struct addrinfo            **result;
-    /* result is a synthetized result, don't call freeaddrinfo on it */
-    struct mk_list               _head;
-};
-
-#define FLB_DNS_LOOKUP_CONTEXT_FOR_EVENT(event) \
-    ((struct flb_dns_lookup_context *) \
-        &((uint8_t *) event)[-offsetof(struct flb_dns_lookup_context, response_event)])
 
 #define FLB_DNS_LEGACY  'L'
 #define FLB_DNS_ASYNC   'A'
